moved error codes module into HyperSearch dir
