+++ nupic/swarming/HypersearchWorker.py
@@ -37,7 +37,7 @@ from nupic.support import clippedObj
 from nupic.support import initLogging
 from nupic.support.configuration import Configuration
 from nupic.support.ExtendedLogger import ExtendedLogger
-from nupic.support.errorcodes import ErrorCodes
+from nupic.swarming.HyperSearch.errorcodes import ErrorCodes
 from nupic.database.ClientJobsDAO import ClientJobsDAO
 from HypersearchV2 import HypersearchV2
